wsd: move parseUri to Uri.hpp
