+++ net/NetUtil.hpp
@@ -110,19 +110,6 @@ void asyncConnect(const std::string& host, const std::string& port, bool isSSL,
 std::shared_ptr<StreamSocket>
 connect(std::string uri, const std::shared_ptr<ProtocolHandlerInterface>& protocolHandler);
 
-/// Decomposes a URI into its components.
-/// Returns true if parsing was successful.
-bool parseUri(std::string uri, std::string& scheme, std::string& host, std::string& port,
-              std::string& pathAndQuery);
-
-/// Decomposes a URI into its components.
-/// Returns true if parsing was successful.
-inline bool parseUri(std::string uri, std::string& scheme, std::string& host, std::string& port)
-{
-    std::string pathAndQuery;
-    return parseUri(std::move(uri), scheme, host, port, pathAndQuery);
-}
-
 inline std::string getDefaultPortForScheme(const std::string& scheme)
 {
     if (scheme == "https://" || scheme == "wss://")
@@ -136,28 +123,6 @@ inline std::string getDefaultPortForScheme(const std::string& scheme)
 // default port numbers are considered equivalent if explicitly included in the
 // compared against peer.
 bool sameOrigin(const std::string& expectedOrigin, const std::string& actualOrigin);
-
-/// Return the locator given a URI.
-inline std::string parseUrl(const std::string& uri)
-{
-    auto itScheme = uri.find("://");
-    if (itScheme != uri.npos)
-    {
-        itScheme += 3; // Skip it.
-    }
-    else
-    {
-        itScheme = 0;
-    }
-
-    const auto itUrl = uri.find('/', itScheme);
-    if (itUrl != uri.npos)
-    {
-        return uri.substr(itUrl); // Including the first slash.
-    }
-
-    return std::string();
-}
 
 } // namespace net
